+++ src/var.h
@@ -23,6 +23,8 @@ int yed_get_default_scroll_offset(void);
 #define DEFAULT_STATUS_LINE_CENTER ""
 #define DEFAULT_STATUS_LINE_RIGHT  "(%p%%)  %l :: %c  %t "
 
+#define DEFAULT_SYNTAX_MAX_LINE_LENGTH 1000
+
 int yed_var_is_truthy(const char *var);
 int yed_get_var_as_int(const char *var, int *out);